+++ gulpfile.js
@@ -3,6 +3,8 @@ var gulpif = require('gulp-if');
 var autoprefixer = require('gulp-autoprefixer');
 var sass = require('gulp-sass');
 var plumber = require('gulp-plumber');
+var cssnano = require('gulp-cssnano');
+var runSequence = require('run-sequence');
 
 gulp.task('sass', function() {
   return gulp.src('scss/nhsuk/nhsuk.scss')
@@ -26,9 +28,20 @@ gulp.task('autoprefix', () =>
 		.pipe(gulp.dest('public/css'))
 );
 
+gulp.task('minify', function() {
+	gulp.src('public/css/nhsuk.css')
+      .pipe(cssnano())
+      .pipe(gulp.dest('public/css'))
+});
+
 gulp.task('watch', function() {
   gulp.watch('scss/nhsuk/**/*.scss', ['sass']);
 });
 
-gulp.task('build', ['sass', 'autoprefix']);
+gulp.task('build', function (callback) {
+  runSequence('sass','autoprefix', 'minify',
+    callback
+  )
+})
+
 gulp.task('default', ['build', 'watch']);